Updated openstack/openstack Project: openstack/openstack-manuals  0786d8e0f93f2999da9535ef64f19d45b9292964
[admin-guide] Document CPU topology configuration

While the configuration of how an instance is placed on the host is
documented, the specifics of how the inner topology of this instance can
be customized is not. Close this gap.

This includes an additional section to the introduction that details
what the terms "socket", "core" and "thread" mean in the context of
OpenStack.

Change-Id: Ia53bd81682f474271d284b3715887f42b29cf2b9