Updated openstack/openstack Project: openstack-infra/project-config  f45b5f55c6167883381a2b5f5677da32ca7563d8
Nodepool elements: Add a script to initialize urandom

In our Xenial images, we see unbound take a while to start because
it uses openssl which uses the getrandom call which can block during
early boot if the nonblocking random number generator is not yet
initialized.

This script uses haveged to quickly initialize the generator.

This commit only includes the script, a later commit will add the
rest of the necessary install steps to the element.

Change-Id: I09d18a0bad6c380fd149660ebfdaf6c12730dc74