Updated openstack/openstack Project: openstack/openstack-manuals  960b15b072bdaf65709d5b961b7a4b184dd55528
Replaced the term "tenant" with "project"

This patch avoids the mismatch between the legacy term tenant and
the future term project by replacing the term tenant with project.

This is a partial fix and the further patches will completely
remove the term "tenant" from the documentation to avoid any kind
of confusion.

Change-Id: I52adfa5831f504145c2f6f0018062c818518633f
Partial-Bug: #1475005