Updated openstack/openstack Project: openstack/horizon  8194f6438a094cafedffc4b78ea368569a61ccb0
Remove embedded CSS

As follow on work to restoring the old network topology, this patch
separates embedded CSS in one of the html files into a separate SCSS
file.

This change is not intended to make topology themable.  There is still
much work that needs to be done and will follow up soon.

Change-Id: I46b7152a24fc1f7a26a1752ba22d80e31ff1ffdb