Updated openstack/openstack Project: openstack/neutron  129fffd5da9aca48f6e7c4afaa52b914379fec9d
Make callback manager Object Oriented friendly

The callback manager was indexing callbacks based on
a callback ID generated from oslo utils reflection.
This presented two problems.

The first was that in py34 get_callable_name would use
__qualname__ which returns the class name the function
is defined on rather than the class of the object itself.
So two classes defined in the same module inheriting from
the same parent class could not both subscribe a method
defined on the parent.

The second more general problem is that two objects which
are instances of the same class cannot subscribe the same
method because they have the same ID.

This adds the hash of the method to the ID to prevent these
issues. The hash by itself could have been used but it's not
very user-friendly so the name is left on for nice log
messages.

Change-Id: Iff1ca8c4ddb58ca5907d21fa0de7f0f292b6fc0e