Updated openstack/openstack Project: openstack/nova  217c61e20aa699c5f1bff6ce5867811c7dd3cb2e
In placement API send microversion header when error

If there is an exception in the placement API application we still
need to send the openstack-api-version header in the response:
errors are versioned too. Therefore catch the exception, add to its
headers and then reraise the exception.

This was a problem identified in the review of the initial framing
of the placement API but left for a later patch. This is that later
patch.

Change-Id: Ib83acbf74907fd9dcbdd4b0a9271ad4bae1c21b1
Partially-Implements: blueprint generic-resource-pools