Updated openstack/openstack Project: openstack/heat  d016928efd5b79ba092c97c7d2b283974ed21a10
Validate the input/output configs in Software Config

When the only way to define a Software Config was via a Heat resource, the
input and output configs were validated by the properties of the resource.
However, subsequently a REST API to create Software Configs directly was
added. That means that configs created in this way do not have the contents
of the inputs and outputs sections validated. This change adds validation
to ensure that the configs always follow the correct schema.

Change-Id: I8c66bb82484b75723524959be753a4cd20c0f84d