Updated openstack/openstack Project: openstack/tempest  c5630c6ac71d7701cb13eb7bde155a1b0cebb5c4
Remove the test_upload_large_objects negative test

The test_upload_large_objects test that is part of the
test_account_quotas_negative.py has been skipped for 2 years now,
and there is no way to run this test in a reliable way since all the
container and account updates are done by background daemons running
in the cluster somewhere.
This commit removes the test from Tempest since it is a test that
won't be run anymore.

Change-Id: Id3917003f0cf3c0ab652c9ea16d22c696f52933f
Closes-Bug: 1310597