Updated openstack/openstack Project: openstack/nova  748ccf07602042a2e0582403d5d9be3ecb2c1d27
rt: Refactor unit test for trackable migrations

There was a functional test in test_resource_tracker.py call
test_no_move_type() that was setting up a resource tracker instance and
populating it only to test whether the migration's migration_type was
one of three that are tracked by the resource tracker.

This was silly, so I converted this into a real unit test, derived from
NoDBTestCase, in test_tracker.py and converted the
_is_trackable_migration() method on ResourceTracker to just be a
module-level method since it wasn't referencing self or anything.

Also added a unit test to ensure that when
_update_usage_from_migration(), is called with a non-tracked migration
(i.e. a live migration), we short-circuit and don't do anything.

Change-Id: I8a8dea0c8e2bcc689a7141bca53a1564325939bc