Updated openstack/openstack Project: openstack/tempest  a6d2b26e145ab42bcd8fbe49d88ee52772df2ca8
Remove base_quotas_client for the reability

The service clients for the volume service are different from the
other services and it is a little difficult to read and debug because
that consists of multiple layers for v1 and v2.
That could be useful for the code optimization, but actually the merit
seems small and unreadable.
This patch removes these layers from the volume clients.

Partially implements blueprint consistent-service-method-names

Change-Id: Ie6d37a72158e4d804727fa756018a37358c1fd59