Updated openstack/openstack Project: openstack/cinder  3ad7384913546a71b32b7e321c035183eedfc255
Specify key algorithm and size for create_key

Fixes a bug that was allowing the key manager to create the default
size and algorithm of a key every time. This patch passes a key size
and algorithm to the create_key command to provide more accurate
information to the key manager.

Another fix will be needed to address the use case of aes-xts with a
key size of 512 -- key managers may not be able to create 512 bit AES
keys.

Change-Id: I166a7685914e25dff9268c86383ad68c41a21530
Closes-Bug: #1514546