Updated openstack/openstack Project: openstack-infra/project-config  f60cc35d7c5babd665e33928937f8adfcd6dd628
Move networking-ovn grenade CI job from experimental to check

Grenade job for OVN project is stable after the changes in
https://review.openstack.org/#/c/343591/ got merged. The same can be
moved to the 'check' queue so that we can monitor the status of the migration
for every patch and fix it whenever it's broken.

This is still a non-voting job.

Change-Id: I73a03e7dacfecd3d270e854326361604bf75e9e7