Updated openstack/openstack Project: openstack/nova  69509889555efa01ff16a2635c97297bc736f7c7
Fix migration list + MigrationList operation

Fixes up the get_migrations cells method and corresponding tests. The
tests were not in line with reality since the MigrationList object was
introduced.

This patch removes the [] + MigrationList operation that was occurring
since the operation is poorly defined and only works by accident. Future
objects work may break this so it should be removed now.

Change-Id: I5dabd8c745cc287322a1ef0f0dd2d4733484188d