Updated openstack/openstack Project: openstack/python-novaclient  38073cab51821e4c29a47bcc7cb33d45b112d1c3
Make wait_for_server_os_boot wait longer

When running the trigger crash dump functional tests on a
local devstack VM running with 4VCPUs and 8GB RAM, I'm getting
timeouts after 60s waiting for the guest OS to boot. The gate
jobs use VMs with 8VCPUs and 8GB RAM, which is fine, but we
shouldn't require people to have a VM that size just to run
novaclient functional tests reliably.

While debugging, I verified that the guest OS is booting, it's
just not done in 60 seconds.

This change bumps the timeout to 300 seconds, which is the
default that Tempest uses for waiting for a server to complete
building. It also adds the last console output from the guest
to the error message for debugging if/when this fails.

Change-Id: I36d43c6e87839856740b2317ee57858ce6357051
Closes-Bug: #1613454