Updated openstack/openstack Project: openstack/nova  33a50eace0f9ae487453135f6ff644e95d577293
Refactor instance action notification sample test

Every instance action notification sample test needs an instance to
work with. Until now each test booted a new instance. This patch
refactors the tests to use a common instance where it is possible.

This optimization speeds up the test execution significantly.
Before the commit:
  Sum of execute time for each test: 151.7733 sec.

After the commit:
  Sum of execute time for each test: 74.7535 sec.

Please note that only 6 actions are tested right now but there will
be ~40 when every notification is transformed so this will give use
speed in the future as well.

Change-Id: I83af64689c26fb8b825c0f1abf81821bbaacb2ac