Updated openstack/openstack Project: openstack-infra/storyboard-webclient  cfc321118516ed199681301836c03d25425d4260
Rework the task list layout

This commit reworks the task list layout so that tasks are arranged
by project and branch. The rows in the list are also modified to make
space for the addition of new fields in future. One effect of this is
that task notes can now be viewed and edited inline.

Change-Id: I8694c906c0a5f3567331574ec3a516dec586414a