Updated openstack/openstack Project: openstack/nova  5e62bd3af6eed8ad9903e229215e62eb59dd7329
Move JSON linting to pep8

The docs job is not run in the gate, add the linting command from
the docs environment that checks for valid JSON files to the pep8
environement so that jobs run it.
Add a simple comment to both invocations to explain what it does.

Note that pep8 is our general linting target and thus should include
these kind of tests.

Change-Id: Ibb2ad29ca59b86fef41ebfe0ec3b8e8ea38e6d72