Updated openstack/openstack Project: openstack/tripleo-heat-templates  e59cef59b66f16338084213277cff6f61867efbf
Add management network to list of networks to ping

Add the end of overcloud deployment, a list of network IP addresses
for isolated networks is used to check connectivity.  The management
network IP address is missing from this list.  This fix adds it in
so that the management network will be checked if it is configured.

Change-Id: I02b020681bed7fbd6d05416280d51a2cbcc50eee
Closes-Bug: 1609554