Updated openstack/openstack Project: openstack/horizon  73630ca4afc303a38679795132a96cf014a94e00
Fix the loss of JSON types when using multipart/form-data

To pass a binary blob in a POST request, browser sets the header
'Content-Type: multipart/form-data', which in turn causes all form
fields' values to be passed as strings. Circumvent this by storing
original field values as a JSON string on client-side and decoding it
on server-side. As a result the setting HORIZON_IMAGES_UPLOAD_MODE =
'legacy' will start working together with Glance V2.

Closes-Bug: #1613703
Change-Id: I53a8fbba15e4c3c6c17d6ef1ffe701634efda149