Updated openstack/openstack Project: openstack/heat  34a5d15ff60a05a175321bc1a1587081edd1588a
Deal with cinder-backup service absent

cinder-backup service is optional, and currently not discoverable with
non-admin scoped tokens.

This patch introduces a new config option `volumes.backups_enabled`
(defaults to True for backward compat).
When set to False, it will fail validation of stacks containing volume
resources with deletion_policy set to Snapshot,
as such stacks will be undeletable in the absence of cinder-backup
service.

Change-Id: I10c4787870e7b240d775aa5d66eb1d0a6d837d8d
Partial-Bug: #1595159