Updated openstack/openstack Project: openstack-infra/nodepool  36484820fd69ed85f78ff6ef3a120ee62786fadb
Stop running DIB usage report

The DIB disk usage reporting is buggy and can result in failed builds.
This option can and should be set on a per image basis using our config
instead of being unconditionally on. This will allow us to workaround
the bug in DIB until it is corrected.

Change-Id: I0c2d901d4e57614c8be690d6ccc025b116de0838