Updated openstack/openstack Project: openstack/python-openstackclient  8cef12ce0847cdc7e95922d0c759d583feb46735
Add shell integration test

These run next to unit tests, but unlike unit tests, they test the
dependent libraries also.  This is to detect incompatible breakage
in those dependencies.  The tests provide CLI-level input and
verify the API calls being made via requests using requests_mock
so the entire stack is tested.

It is possible we want to run these separate from the unit tests.
They do not belong in the functional tests as they do not require
a functional cloud for testing.

Depends-on: I426548376fc7d3cdb36501310dafd8c44d22ae30
Change-Id: I356956fcc4ff35191a6ad6a085b75cf370434b09