Updated openstack/openstack Project: openstack/keystone  f5468986263e6dc4fd45b49e99ff2951e9f2cd82
Removes a redundant test from FernetAuthWithTrust

It turns out that the behaviour tests by
test_trust_get_token_fails_if_trustee_disabled is not actually different
between Fernet and UUID if the tokens are properly revoked. The original
test didn't ensure that any time had passed between creating and revoking
a token. This adds a new test showing the different behaviour and modifies
the existing test to works.

Previously this test would fail intermittently when hitting a second
boundary.

If the creation and revocation happen in the same second then Fernet will
indeed respond with a Forbidden. If the revocation happens at least one
second after it will response with the expected Unauthorized.

Change-Id: I31e55e92cd745c34254e96491f97112ff0513baf