Updated openstack/openstack Project: openstack/ironic  a1ffa8ff91d34d32fed3fe8c4571b1576e5a3e6e
DevStack: No need to change the ramdisk filesystem type

This patch is removing the conditional that was changing the ramdisk
fileystem type from tmpfs to ramfs in case the amount of RAM was lower
than 1GB.

That code was used before to make the bash ramdisk to boot on machines
with less than 1GB of ram, now we do have TinyIPA so I believe this code
is not needed any more.

Ramfs is old and should be replaced with tmpfs which overcomes many of
it's limitation.

Change-Id: I54fd06213e003aba705118ea0cff72f889cfcc08