Updated openstack/openstack Project: openstack-infra/system-config  4eefe525254d46fb01ef6c9e1ba713c8efd22521
Update logrotate for mirror server

Since this is currently unmanaged by openstack-infra, the default is
to rotate logs every week, 52 times.  Needless to say, we are running
out of HDD space.  Now we rotate every day, 7 times.  This will cut
down on the amount of HDD space apache2 logs require.

Change-Id: I7baf62a36432f5d322e7f63e8c89bd6205974ca8
Signed-off-by: Paul Belanger <pabelanger@redhat.com>