Updated openstack/openstack Project: openstack/heat  222d70cc9b9023e048b30bcb7f214f3547e6916b
api-ref doc for stack (2)

This patch continues the fix to the api-ref doc for stack resource. To
ensure upcoming patches are not messing up the parameter defintions,
this patch is cleansing up the parameters that appear both in URLs and
request/response bodies: 'stack_id', 'stack_name', 'resource_name',
'event_id', 'snapshot_id', etc. These parameters sometimes are playing
different roles although the name is the same.
Other fixes included in this patch:

- added back the stack-outputs and stack-snapshots into index
- split resource-types out of stack-templates and add reference to it
  into index (git misunderstood this as a renaming, :))

Change-Id: I2b486a9b85d23d3225a5701c2265ecea2f4712f7