Updated openstack/openstack Project: openstack-infra/project-config  c1a6c49e3591e5d28e3b4f4df452dee6d55df497
Conditionally run gate-tempest-dsvm-neutron-linuxbridge in nova check

gate-tempest-dsvm-neutron-linuxbridge is gating for neutron which
means nova can merge something to break it and block neutron from
landing any code. In an effort to ease into gating nova changes
on gate-tempest-dsvm-neutron-linuxbridge we start with only running
it on changes to nova/network/* files. We can increase the exposure
as necessary with the ultimate goal to be fully gating when we
start dropping nova-network-backed jobs.

This is intentionally only the check queue for now since we (a) think
the risk exposure for a co-gating failure is low and (b) we want to
avoid hammering the CI system with more jobs if we can help it.

Change-Id: I21af439a90104960bcb1cfbcb8afeb909f96a2ad
Related-Bug: #1612281