Updated openstack/openstack Project: openstack/nova  a0ad1258a537adab8f533240ffa2f26671af68dd
Add support for inventories to placement API

GET to list /resource_providers/{uuid}/inventories
POST to create a single new inventory via
    /resource_provider/{uuid}/inventories
PUT to (re-)set all inventories at
    /resource_providers/{uuid}/inventories
GET, DELETE or UPDATE one inventory of a particular resource class
   /resource_provider/{uuid}/inventories/{resource_class}

Inventories are presented with a view marker named
'resource_provider_generation' which is used to enable consistent
views of inventory across concurrent updates. They are also used
in fashion similar to ETags to avoid concurrent updates.

ETags were considered for this functionality but discussion
amongst interested parties led to the conclusion that the marker
was more explicit, especially in the case of taking an inventory
representation "seen" in a collection representation, updating it,
and sending it back to the server.

Change-Id: Ib49c7cddd2f15869f01e8b1e09d8378c26fb7ddc
Partially-Implements: blueprint generic-resource-pools