Updated openstack/openstack Project: openstack/api-site  f8546708fabd5d75a091c537309cdf5439a1bbda
Identity v2.0 APIs are deprecated

Review https://review.openstack.org/#/c/351370/ offically marked
the Identity v2.0 APIs deprecated within keystone, but the follow-on
table of contents that references it was not updated. So one shows
CURRENT and the followed link shows DEPRECATED.

This patch updates the table of contents to properly show DEPRECATED
for the v2.0 APIs.

For comparision, see the current version of this page:
http://developer.openstack.org/api-ref.html

Change-Id: Ia56829541a27896b588d296889141d5059570320