Updated openstack/openstack Project: openstack-infra/project-config  945b10084d91485ea73fe9ce3f6e788f1f4dd069
Ensure ntpdate is on our test images

We want to be able to set the time in big steps at the beginning of test
runs and one option for doing so is with ntpdate. Ensure ntpdate is
installed on all our images by putting it into the infra package needs
element.

Note that the package name appears to be the same across ubuntu, centos,
and fedora.

Change-Id: Ib3fd4afe5a89d8a799cc15c57254aaf11b6aa3e5