Updated openstack/openstack Project: openstack/python-novaclient  296a3c0519b6fb269af616ad4c4910ad85ad5ac5
Deprecate network-* commands and clamp to microversion 2.35

This introduces a helper to clamp the client microversion to 2.35,
which is the last version to support the network proxy. We print a
deprecation warning if those commands are used, and mark them as
deprecated in the help text.

This is a network-specific user-friendly bit of sugar to make sure
that nova-network users aren't cut out before we actually drop the
support for it on the server side.

Note that quotas and limits are special because only the network
related resources in those are not returned with 2.36. So this
change handles 2.36 separately for quota-update and
quota-class-update, and deprecates the network resource quota
update arguments for <2.35 as an indication those are going away.

As expected, several of the functional tests have to be updated
to work with the new world that is microversion 2.36.

Related to blueprint deprecate-api-proxies

Co-Authored-By: Matt Riedemann <mriedem@us.ibm.com>

Change-Id: Id68c2dbef29b201aa7c8ef9417432feb5596529a