Updated openstack/openstack Project: openstack/neutron  821ddd36fb6f3fee257ceedcf6f02d0635d5eba7
Check the router gateway IPs prefixlen existence

When we remove the router gateway, if it has multiple IPs
for the gateway port, the router remove procedure will not
work properly and will cause a router remove infinite loop.
This is because the neutron server only gives the first (lowest)
gateway IP a 'prefixlen', and not the secondary IPs, so
the router remove will get a KeyError.

This patch adds a `prefixlen` check before removing the gateway IPs
conntrack state.

Change-Id: I425897578d8bd44f4e5f9dd44de9717ae4e50fcd
Closes-Bug: #1609248