Updated openstack/openstack Project: openstack-infra/puppet-mediawiki  28767a548629cbc929414a43769f166bb55eff61
Switch from old recaptcha to recaptcha-nocaptcha

The newer recaptcha-nocaptcha implementation is a little better at
thwarting script-driven spammers and vandals, so switch to it. Leave
the parameter names for the old recaptcha implementation in place
until our consumer in the openstack_project::wiki class no longer
attempts to set it.

Change-Id: I435484cd65a028d774dfd920ca19f3077d4e03fb