Updated openstack/openstack Project: openstack-infra/shade  c40cc1938185ad80cc2f62ca4f6e8665a11a6fbf
Use cloud fixtures from the unittest base class

We need to do things to protect against local environment leaking into
the test fixtures when we create the cloud objects. Those things are all
done in the base unittest TestCase class, but for reasons that are hard
to fathom, we ignore that and create our own clouds in many of the
unitttest classes. This leads to problems if a user running the
unittests has cache config information in their local clouds.yaml.

Fix it.

Change-Id: I022d541d8e98bf4b6691bf0a91e3b7d20b2b7456