Updated openstack/openstack Project: openstack-infra/shade  aa6505af09e2fe69e961cefb1e8af0b751f03455
Update HACKING.rst with a couple of shade specific notes

Reasonable people disagree on finer points of style. However, it's
pretty well understood that being consistent within a single codebase is
more important that specific choices globally.

There are a few things worth pointing out that are true in the context
of shade which may not be true in other places, so active communication
is likely the best choice here.

Change-Id: Ib1ceb5d6f51f84fa4bc40e68e9e231a60138507d