Updated openstack/openstack Project: openstack/python-novaclient  1b6cc4a6bc48ef4c049c16af6e64f8da5875fc86
Cap image API deprecated methods at 2.35

The image proxy API GET/DELETE methods are deprecated
at microversion 2.36 and will return a 404 after that.

This adds the wraps decorator to those python API methods
so set the cap at 2.35 so rather than get a 404 you'll
get a more uesful VersionNotFoundForAPIMethod.

Note that set_meta and delete_meta are not decorated
because we missed deprecating the image-metadata proxy API
with the 2.36 microversion. That will be fixed in Ocata.

Related to blueprint deprecate-api-proxies

Change-Id: Ie65efadb5c65e8a624ffd0315a634accd49f1c30