Updated openstack/openstack Project: openstack-infra/shade  c9851b92af78001171c6637048c4f954e7e96297
Stop getting extra flavor specs where they're useless

We supplement flavors with the values in extra_specs as a default.
However, there are several places where we're getting flavor objects on
behalf of the user where we know that we will not use them. Make sure
that we can plumb through the get_extra flag, and then pass it as False
in the places where getting the extra specs is just a waste of time.

Change-Id: I46d8ed5537a94a0440dd42d305221686cf1b225b