Updated openstack/openstack Project: openstack/heat  d1cb214a1c0a678f0e9cb731d3b57084f5a10c91
Depend on tempest in test-requirements.txt

This replaces the requirements.stub dependency approach for
heat_integrationtests with a single dependency on tempest in
test-requirements.txt.

tempest provides all the required dependencies for the current
heat_integrationtests, and this helps the transition of being able to
run tests both with the current runner and as a tempest plugin.

blueprint tempest-plugin-support

Change-Id: I6fc5886feec5de71826f640cd29990114bd7781c