Updated openstack/openstack Project: openstack/swift  d98928caa99cbfac1b8710ac0761925a8146422e
py3: tox.ini: use substituation to py35 commands

Use tox substitution to avoid duplicating testenv:py34 commands in
testenv:py35, to not have to maintain the whitelist of Python 3 tests
in two different places.

Write also the list of tests in a newline to be able to more easily
add new unit tests.

Change-Id: I6e7f238f1c5d3fc9b6560918dcbb93e9dd8ec084