Updated openstack/openstack Project: openstack-infra/release-tools  3f615682a11df18f38837c4f3fd1a0c72b25e057
Authenticate before doing group membership tests

Group membership tests (used to determine if a PUT/DELETE call
was superfluous or not) were not using authentication and
therefore always returned 403 (which resulted in PUT requests
to always be needed).

Use authentication in group membership tests, and test for a more
precise return code.

Change-Id: I7784ea7334953c333744c603e0719187cac54951