Updated openstack/openstack Project: openstack/tripleo-heat-templates  072404b5693439b728d49d26c2c11ed69172a40d
Don't trigger mariadb upgrade dump/restore when not needed

When upgrading from mariadb X.Y.Z to mariadb X.Y.Ž (X.Y part stays the
same), the dump/restore of mariadb shouldn't be necessary. Therefore we
now only check for up to the first 2 fields of the version string when
determining if we should trigger the dump/restore operation.

Closes-Bug: #1615721
Change-Id: Ib7af8bfb121f5c83184d51b3c6dc657108c25973