Updated openstack/openstack Project: openstack-infra/system-config  dbb556d2df20f61c90b1f7f3b1df0841f9f851d0
Set up HTTPS for releases.openstack.org site

Add separate HTTPS cert/key/chain files for the
releases.openstack.org site since our current certificate vendor
requires us to renew a cert to update the SubjectAltNames list in
it. Note that this depends on RFC 4366 (Server Name Indication)
support in browsers, but basically all "modern" browsers should be
fine with it (if someone is still using IE6, I'll be happy to print
out and mail them a copy of the site). Also, the relatively strict
TLS options we've been using for the other HTTPS sites on this
server basically already preclude browsers of that vintage, so it's
unlikely to come up at all.

This site can be folded into the static.openstack.org multi-site
cert when it's renewed next year and switched to the normal
static-https-redirect.vhost.erb template at that time.

Change-Id: I92f7fd163fc36ab06116233622a9a07c5f20440d